Fix nos tipos de alguns campos de US e suas sub-classes

+++ src/main/java/com/ufcg/si1/model/unidadesaude/PostoSaude.java
@@ -15,7 +15,7 @@ public class PostoSaude extends UnidadeSaude {
     }
 
     public PostoSaude(String descricao, Set<Especialidade> especialidades, int quantidadeMedicos,
-                      int taxaDiariaAtendimentos) {
+                      float taxaDiariaAtendimentos) {
         super(descricao, especialidades, quantidadeMedicos);
         this.taxaDiariaAtendimentos = taxaDiariaAtendimentos;
     }
